Redesign the local provider logic to support instance types and go with simulator soly
