+++ example/cluster/local/example-cloud-simulator-config.yaml
@@ -1,16 +1,19 @@
-head_node:
-    ip: your_head_ip
 
-    # Specify the resources of the node.
-    # We assume all the worker nodes are the same with head node
-    CPU: number-of-cores
-    memory: size-in-mb
+# Define one or more instance types with the information of its hardware resources
+# Then you specify the instance type for each node in the node list
+instance_types:
+    your_instance_type:
+        # Specify the resources of this instance type.
+        resources:
+            CPU: number-of-cores
+            memoryMb: size-in-mb
 
-    # You may need to supply a public ip for the head node if you need
-    # to start cluster from outside of the cluster's network
-    # (e.g. the cluster is in an AWS VPC and you're starting the cluster from your laptop)
-    # This is useful when debugging the local node provider with cloud VMs.
-    # external_ip: your_head_public_ip
-worker_nodes:
-    - ip: work_node_1_ip
-    - ip: work_node_2_ip
+# List of nodes with the ip and its node type defined in the above list
+nodes:
+    - ip: node_1_ip
+      node_type: your_instance_type
+      # You may need to supply a public ip for the head node if you need
+      # to start cluster from outside of the cluster's network
+      # external_ip: your_head_public_ip
+    - ip: node_2_ip
+      node_type: your_instance_type
