--- example/cluster/local/example-docker.yaml
@@ -1,35 +1,0 @@
-# A unique identifier for the cluster.
-cluster_name: example-docker
-
-# Enable container
-docker:
-    enabled: True
-
-# Cloud-provider specific configuration.
-provider:
-    type: local
-    head_node:
-        ip: your_head_ip
-
-        # Specify the resources of the node.
-        # We assume all the worker nodes are the same with head node
-        CPU: number-of-cores
-        memory: size-in-mb
-
-        # You may need to supply a public ip for the head node if you need
-        # to start cluster from outside of the cluster's network
-        # (e.g. the cluster is in an AWS VPC and you're starting the cluster from your laptop)
-        # This is useful when debugging the local node provider with cloud VMs.
-        # external_ip: your_head_public_ip
-    worker_nodes:
-        - ip: your-work-node-1-ip
-        - ip: your-work-node-2-ip
-    # If you are using a Cloud Simulator, you can specify the simulator address.
-    # Instead of specify the head_node and worker_nodes
-    # cloud_simulator_address: 172.31.6.253:8100
-auth:
-    ssh_user: ubuntu
-    # Specify the private key file for login to the nodes
-    ssh_private_key: ~/.ssh/your-private-key.pem
-    # Set proxy if you are in corporation network. For example,
-    # ssh_proxy_command: "ncat --proxy-type socks5 --proxy your_proxy_host:your_proxy_port %h %p"
