+++ python/cloudtik/providers/local/defaults.yaml
@@ -3,24 +3,9 @@ from: defaults
 
 provider:
     type: local
-    head_node:
-        ip: your_head_ip
 
-        # Specify the resources of the node.
-        # We assume all the worker nodes are the same with head node
-        # CPU: number-of-cores
-        # memory: size-in-mb
-
-        # You may need to supply a public ip for the head node if you need
-        # to start cluster from outside of the cluster's network
-        # (e.g. the cluster is in an AWS VPC and you're starting the cluster from your laptop)
-        # This is useful when debugging the local node provider with cloud VMs.
-        # external_ip: your_head_public_ip
-    worker_nodes: []
-        # - ip: work_node_1_ip
-        # - ip: work_node_2_ip
-    # Optional when running automatic cluster management on prem. If you use a cloud simulator,
-    # then you can launch multiple clusters on the same set of machines, and the cloud simulator
+    # We need to use Cloud Simulator for the best local cluster management
+    # You can launch multiple clusters on the same set of machines, and the cloud simulator
     # will assign individual nodes to clusters as needed.
     # cloud_simulator_address: "<host>:<port>"
 
@@ -29,3 +14,14 @@ auth:
     ssh_user: ubuntu
     # Optional if an ssh private key is necessary to ssh to the cluster.
     # ssh_private_key: ~/.ssh/id_rsa
+
+available_node_types:
+    head.default:
+        node_config:
+            instance_type: default.instance.type
+    worker.default:
+        node_config:
+            instance_type: default.instance.type
+
+# Specify the node type of the head node (as configured above).
+head_node_type: head.default
